// frontend/types/documentTemplates.ts
export type DocumentTemplateKey =
    | 'finance_statement_t49'
    | 'finance_payslip'
    | 'finance_timesheet_t13'
    | 'order_invoice'
    | 'order_invoice_alt'
    | 'order_supply_contract'
    | 'order_supply_specification'
    | 'order_service_contract'
    | 'order_work_contract'
    | 'order_outgoing_act'
    | 'purchase_invoice'
    | 'purchase_upd_status_1'
    | 'purchase_upd_status_2'
    | 'purchase_torg_12'
    | 'shipment_upd_status_1'
    | 'shipment_upd_status_2'
    | 'shipment_torg_12'
    | 'shipment_transport_waybill';